Implemented an LSTM with the gates specified in the paper
The PyTorch implementation is slightly different with regards
to the biases. Instead of having one bias vector per gates
they use two, one for the input and the other one for the
hidden state. To make sure we are doing the same thing as in the
paper this is a straightforward implementation of the equations
11 to 15. This is not optimized.